Modified to use resp charge. Tested on charged system, acetonitrile solvent.
Committer: Fang Liu <fangliu@comet-ln2.sdsc.edu>

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
       modified:   set_up_solvated_AmberMD.py
       modified:   ../devtools/conda-envs/test_env.yaml

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
      modified:   set_up_solvated_AmberMD.py
      modified:   ../devtools/conda-envs/test_env.yaml